Fixed scrollSpy
