+++ themes/cloudposse/src/js/app.js
@@ -6,7 +6,6 @@ import './ui';
 import './featherlight';
 import './expandable';
 import './scroll';
-import './scrollspy';
 import './search';
 
 $(document).ready(function () {
